+++ feature/main/src/main/java/kr/co/main/accountbook/main/AccountBookBottomSheet.kt
@@ -17,36 +17,27 @@ import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDefaults
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.DisplayMode
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.SelectableDates
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.material3.rememberDatePickerState
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.DialogProperties
 import kr.co.domain.entity.AccountBookEntity
+import kr.co.main.accountbook.model.CustomDatePickerDialog
 import kr.co.main.accountbook.model.DatePickerType
 import kr.co.main.accountbook.model.DateRangeOption
 import kr.co.main.accountbook.model.getDisplay
@@ -55,12 +46,8 @@ import kr.co.ui.icon.dreamicon.DatePicker
 import kr.co.ui.theme.Paddings
 import kr.co.ui.theme.colors
 import kr.co.ui.theme.typo
-import java.text.SimpleDateFormat
 import java.time.LocalDate
-import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
-import java.util.Date
-import java.util.Locale
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -147,6 +134,8 @@ internal fun AccountBookCalendarBottomSheet(
     var showDatePicker by remember { mutableStateOf(false) }
     var datePickerType by remember { mutableStateOf(DatePickerType.START) }
 
+    val context = LocalContext.current
+
     ModalBottomSheet(
         onDismissRequest = dismissBottomSheet,
         sheetState = sheetState,
@@ -216,20 +205,34 @@ internal fun AccountBookCalendarBottomSheet(
             )
         }
     }
-
     if (showDatePicker) {
         CustomDatePickerDialog(
-            startDate = currentStartDate,
-            endDate = currentEndDate,
-            datePickerType = datePickerType,
+            date = if (datePickerType == DatePickerType.START) currentStartDate else currentEndDate,
             onClickCancel = { showDatePicker = false },
             onClickConfirm = { selectedDate ->
-                val selectedLocalDate =
-                    LocalDate.parse(selectedDate, DateTimeFormatter.BASIC_ISO_DATE)
+                val formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd")
+                val selectedLocalDate = LocalDate.parse(selectedDate, formatter)
                 if (datePickerType == DatePickerType.START) {
-                    currentStartDate = selectedLocalDate
+                    if (selectedLocalDate.isAfter(currentEndDate) || selectedLocalDate == currentEndDate) {
+                        Toast.makeText(
+                            context,
+                            "시작일이 종료일보다 이후입니다.",
+                            Toast.LENGTH_SHORT
+                        ).show()
+
+                    } else {
+                        currentStartDate = selectedLocalDate
+                    }
                 } else {
-                    currentEndDate = selectedLocalDate
+                    if (selectedLocalDate.isBefore(currentStartDate) || selectedLocalDate == currentStartDate) {
+                        Toast.makeText(
+                            context,
+                            "종료일이 시작일보다 이전입니다.",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    } else {
+                        currentEndDate = selectedLocalDate
+                    }
                 }
                 tempSelectedOption = DateRangeOption.OTHER
                 showDatePicker = false
@@ -356,106 +359,6 @@ private fun OptionBox(
             color = textColor,
             style = MaterialTheme.typography.bodyMedium,
             modifier = Modifier.padding(horizontal = 16.dp)
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun CustomDatePickerDialog(
-    startDate: LocalDate? = LocalDate.now(),
-    endDate: LocalDate? = LocalDate.now(),
-    datePickerType: DatePickerType,
-    onClickCancel: () -> Unit,
-    onClickConfirm: (yyyyMMdd: String) -> Unit
-) {
-    val initialDateMillis = when (datePickerType) {
-        DatePickerType.START -> startDate?.atStartOfDay(ZoneOffset.UTC)?.toInstant()?.toEpochMilli()
-            ?: System.currentTimeMillis()
-
-        DatePickerType.END -> endDate?.atStartOfDay(ZoneOffset.UTC)?.toInstant()?.toEpochMilli()
-            ?: System.currentTimeMillis()
-    }
-
-    val datePickerState = rememberDatePickerState(
-        yearRange = IntRange(2000, 2050),
-        initialDisplayMode = DisplayMode.Picker,
-        initialSelectedDateMillis = initialDateMillis,
-        selectableDates = object : SelectableDates {
-            override fun isSelectableDate(utcTimeMillis: Long): Boolean {
-                return true
-            }
-        }
-    )
-    val initialSelectedDate = remember { datePickerState.selectedDateMillis }
-    val context = LocalContext.current
-
-    LaunchedEffect(datePickerState.selectedDateMillis) {
-        if (initialSelectedDate != datePickerState.selectedDateMillis) {
-            datePickerState.selectedDateMillis?.let { selectedDateMillis ->
-                val date = Date(selectedDateMillis)
-                val formatter = SimpleDateFormat("yyyyMMdd", Locale.KOREA)
-                val formattedDate = formatter.format(date)
-
-                if (datePickerType == DatePickerType.START) {
-                    val selectedStartDate =
-                        LocalDate.parse(formattedDate, DateTimeFormatter.BASIC_ISO_DATE)
-                    if (selectedStartDate.isAfter(endDate) || selectedStartDate == endDate) {
-                        Toast.makeText(
-                            context,
-                            "시작일이 종료일보다 이후입니다.",
-                            Toast.LENGTH_SHORT
-                        ).show()
-                        onClickCancel()
-                        return@LaunchedEffect
-                    }
-                } else {
-                    val selectedEndDate =
-                        LocalDate.parse(formattedDate, DateTimeFormatter.BASIC_ISO_DATE)
-                    if (selectedEndDate.isBefore(startDate) || selectedEndDate == startDate) {
-                        Toast.makeText(
-                            context,
-                            "종료일이 시작일보다 이전입니다.",
-                            Toast.LENGTH_SHORT
-                        ).show()
-                        onClickCancel()
-                        return@LaunchedEffect
-                    }
-                }
-                onClickConfirm(formattedDate)
-            }
-        }
-    }
-
-    DatePickerDialog(
-        onDismissRequest = { onClickCancel() },
-        confirmButton = {},
-        colors = DatePickerDefaults.colors(
-            containerColor = MaterialTheme.colors.white
-        ),
-        shape = RoundedCornerShape(8.dp),
-        properties = DialogProperties(usePlatformDefaultWidth = false)
-    ) {
-        DatePicker(
-            state = datePickerState,
-            colors = DatePickerDefaults.colors(
-                selectedDayContainerColor = MaterialTheme.colors.primary,
-                selectedYearContainerColor = MaterialTheme.colors.primary,
-                containerColor = MaterialTheme.colors.white,
-                dateTextFieldColors = TextFieldDefaults.colors(
-                    unfocusedContainerColor = Color.Transparent,
-                    focusedContainerColor = Color.Transparent,
-                    focusedIndicatorColor = MaterialTheme.colors.primary,
-                    unfocusedIndicatorColor = MaterialTheme.colors.primary,
-                    disabledIndicatorColor = MaterialTheme.colors.primary,
-                    cursorColor = MaterialTheme.colors.primary,
-                    focusedTextColor = MaterialTheme.colors.primary,
-                    errorContainerColor = Color.Transparent,
-                    focusedLabelColor = MaterialTheme.colors.primary,
-                    unfocusedLabelColor = MaterialTheme.colors.primary,
-                    disabledLabelColor = MaterialTheme.colors.primary,
-                ),
-            )
         )
     }
 }
